css upload

+++ src/controllers/videoController.js
@@ -1,39 +1,51 @@
 import Video from "../models/Video"
+import User from "../models/Users"
 
 export const home = async(req, res) => {
     const videos = await Video.find({})
+        .sort({created:"desc"})
+        .populate("owner")
     return res.render("home", {pageTitle:"Home", videos})
 }
 
 export const watch = async(req, res) => {
-    try{
-        const { id } = req.params
-        const video = await Video.findById(id)
-        return res.render("watch", { pageTitle: video.title, video });
-    } catch {
+    const { id } = req.params
+    const video = await Video.findById(id).populate("owner")
+    if(!video){
+        console.log("video not found")
         return res.status(404).render("404", {pageTitle:"Video not found"})
     }
+    return res.render("watch", { pageTitle: video.title, video });
 }
 
 export const getEdit = async(req, res) => {
-    try{
     const { id } = req.params;
+    const {user:{_id}} = req.session
     const video = await Video.findById(id)
-    return res.render("edit", { pageTitle : "Edit", video });
-    } catch {
-        return res.status(404).render("404", {pageTitle:"Video not found"})
+    if(!video){
+        return res.status(400).render("404",{pageTItle:"Video not found."})
     }
+    if(String(video.owner) !== String(_id)){
+        console.log("not match")
+        return res.status(403).redirect("/")
+    }
+    return res.render("edit", { pageTitle : "Edit", video });
 }
 
 export const postEdit = async(req, res) => {
     const { id } = req.params;
+    const {user:{_id}} = req.session
     const { title, description, hashtags } = req.body;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", {pageTItle:"Video not found."})
     }
+    if(String(video.owner) !== String(_id)){
+        console.log("not match")
+        return res.status(403).redirect("/")
+    }
     await Video.findByIdAndUpdate(id, {title, description, hashtags : Video.formatHashtags(hashtags)})
-    return res.redirect(`/videos/${id}`);
+    return res.redirect("/");
 }
 
 export const getUpload = (req, res) => {
@@ -41,24 +53,43 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async(req, res) => {
-    const { title, description, hashtags } = req.body;
+    const {
+        body:{ title, description, hashtags},
+        file:{path:fileUrl},
+        session:{user:{_id}},
+        } = req;
+
     try{
-        await Video.create({
+        const newVideo = await Video.create({
+            fileUrl,
             title,
             description,
+            owner:_id,
             hashtags : Video.formatHashtags(hashtags)
         });
+        const user = await User.findById(_id)
+        user.videos.push(newVideo._id)
+        user.save()
         return res.redirect("/")
     } catch(error){
+            console.log(error)
             return res.status(400).render("upload", { 
-            pageTitle :"Upload Page", 
-            errorMessage:error._message 
+            pageTitle:error, 
         })
     }
 }
 
 export const deleteVideo = async(req, res) => {
     const {id} = req.params;
+    const {user:{_id}} = req.session
+    const video = await Video.findById(id)
+    if(!video){
+        return res.status(400).render("404",{pageTItle:"Video not found"})
+    }
+    if(String(video.owner) !== String(_id)){
+        console.log("not match")
+        return res.status(403).redirect("/")
+    }
     await Video.findByIdAndDelete(id)
     return res.redirect("/")
 }
@@ -71,7 +102,7 @@ export const search = async(req, res) => {
             title: {
                 $regex : new RegExp(`${keyword}$`, "i"),
             }
-        });
+        }).populate("owner")
     }
     return res.render("search", {pageTitle:"Search",  videos})
 }
